Add base class for Poi/Pia payment methods
