+++ Model/PaymentOnInvoicePaymentMethod.php
@@ -9,20 +9,12 @@
 
 namespace Wirecard\Oxid\Model;
 
-use Wirecard\PaymentSdk\Config\Config;
-use Wirecard\PaymentSdk\Config\PaymentMethodConfig;
-use Wirecard\PaymentSdk\Entity\AccountHolder;
-use Wirecard\PaymentSdk\Transaction\PoiPiaTransaction;
-use Wirecard\PaymentSdk\Transaction\Transaction;
-
-use Wirecard\Oxid\Core\Helper;
-
 /**
  * Payment method implementation for Payment on Invoice
  *
  * @since 1.3.0
  */
-class PaymentOnInvoicePaymentMethod extends SepaCreditTransferPaymentMethod
+class PaymentOnInvoicePaymentMethod extends BasePoiPiaPaymentMethod
 {
     /**
      * @inheritdoc
@@ -30,137 +22,4 @@ class PaymentOnInvoicePaymentMethod extends SepaCreditTransferPaymentMethod
      * @since 1.3.0
      */
     protected static $_sName = "paymentoninvoice";
-
-    /**
-     * @inheritdoc
-     *
-     * @var bool
-     *
-     * @since 1.3.0
-     */
-    protected static $_bMerchantOnly = false;
-
-    /**
-     * @inheritdoc
-     *
-     * @return Config
-     *
-     * @since 1.3.0
-     */
-    public function getConfig()
-    {
-        $oConfig = parent::getConfig();
-
-        $oPaymentMethodConfig = new PaymentMethodConfig(
-            PoiPiaTransaction::NAME,
-            $this->_oPayment->oxpayments__wdoxidee_maid->value,
-            $this->_oPayment->oxpayments__wdoxidee_secret->value
-        );
-
-        $oConfig->add($oPaymentMethodConfig);
-
-        return $oConfig;
-    }
-
-    /**
-     * Get the current transaction to be processed
-     *
-     * @return Transaction
-     *
-     * @since 1.3.0
-     */
-    public function getTransaction()
-    {
-        return new PoiPiaTransaction();
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return array
-     *
-     * @since 1.3.0
-     */
-    public function getConfigFields()
-    {
-        $aAdditionalFields = [
-            'descriptor' => [
-                'type'  => 'select',
-                'field' => 'oxpayments__wdoxidee_descriptor',
-                'options' => [
-                    '1' => Helper::translate('wd_yes'),
-                    '0' => Helper::translate('wd_no'),
-                ],
-                'title' => Helper::translate('wd_config_descriptor'),
-                'description' => Helper::translate('wd_config_descriptor_desc'),
-            ],
-            'additionalInfo' => [
-                'type'  => 'select',
-                'field' => 'oxpayments__wdoxidee_additional_info',
-                'options' => [
-                    '1' => Helper::translate('wd_yes'),
-                    '0' => Helper::translate('wd_no'),
-                ],
-                'title' => Helper::translate('wd_config_additional_info'),
-                'description' => Helper::translate('wd_config_additional_info_desc'),
-            ],
-            'deleteCanceledOrder' => [
-                'type' => 'select',
-                'field' => 'oxpayments__wdoxidee_delete_canceled_order',
-                'options' => [
-                    '1' => Helper::translate('wd_yes'),
-                    '0' => Helper::translate('wd_no'),
-                ],
-                'title' => Helper::translate('wd_config_delete_cancel_order'),
-                'description' => Helper::translate('wd_config_delete_cancel_order_desc'),
-            ],
-            'deleteFailedOrder' => [
-                'type' => 'select',
-                'field' => 'oxpayments__wdoxidee_delete_failed_order',
-                'options' => [
-                    '1' => Helper::translate('wd_yes'),
-                    '0' => Helper::translate('wd_no'),
-                ],
-                'title' => Helper::translate('wd_config_delete_failure_order'),
-                'description' => Helper::translate('wd_config_delete_failure_order_desc'),
-            ],
-        ];
-
-        return array_merge(parent::getConfigFields(), $aAdditionalFields);
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return array
-     *
-     * @since 1.3.0
-     */
-    public function getPublicFieldNames()
-    {
-        return array_merge(parent::getPublicFieldNames(), [
-            'descriptor',
-            'additionalInfo',
-            'deleteCanceledOrder',
-            'deleteFailedOrder',
-        ]);
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param PoiPiaTransaction $oTransaction
-     * @param Order             $oOrder
-     *
-     * @since 1.3.0
-     */
-    public function addMandatoryTransactionData(&$oTransaction, $oOrder)
-    {
-        $oAccountHolder = new AccountHolder();
-        $oShop = Helper::getShop();
-        $oAccountHolder->setFirstName($oShop->oxshops__oxfname->value);
-        $oAccountHolder->setLastName($oShop->oxshops__oxlname->value);
-
-        $oTransaction->setAccountHolder($oAccountHolder);
-    }
 }
