+++ Tests/Unit/Model/PaymentOnInvoicePaymentMethodTest.php
@@ -10,7 +10,6 @@
 use Wirecard\Oxid\Model\PaymentOnInvoicePaymentMethod;
 
 use Wirecard\PaymentSdk\Config\Config;
-use Wirecard\PaymentSdk\Transaction\PoiPiaTransaction;
 
 class PaymentOnInvoicePaymentMethodTest extends OxidEsales\TestingLibrary\UnitTestCase
 {
@@ -23,18 +22,6 @@ class PaymentOnInvoicePaymentMethodTest extends OxidEsales\TestingLibrary\UnitTe
     {
         parent::setUp();
         $this->_oPaymentMethod = new PaymentOnInvoicePaymentMethod();
-    }
-
-    public function testGetConfig()
-    {
-        $oConfig = $this->_oPaymentMethod->getConfig();
-        $this->assertInstanceOf(Config::class, $oConfig);
-    }
-
-    public function testGetTransaction()
-    {
-        $oTransaction = $this->_oPaymentMethod->getTransaction();
-        $this->assertInstanceOf(PoiPiaTransaction::class, $oTransaction);
     }
 
     /**
@@ -52,36 +39,5 @@ class PaymentOnInvoicePaymentMethodTest extends OxidEsales\TestingLibrary\UnitTe
             'for oxid' => [true, 'wdpaymentoninvoice'],
             'not for oxid' => [false, 'paymentoninvoice'],
         ];
-    }
-
-    public function testGetConfigFields()
-    {
-        $aFields = $this->_oPaymentMethod->getConfigFields();
-        $this->assertEquals([
-            'apiUrl',
-            'httpUser',
-            'httpPassword',
-            'testCredentials',
-            'maid',
-            'secret',
-            'descriptor',
-            'additionalInfo',
-            'deleteCanceledOrder',
-            'deleteFailedOrder',
-        ], array_keys($aFields));
-    }
-
-    public function testGetPublicFieldNames()
-    {
-        $aPublicFieldNames = $this->_oPaymentMethod->getPublicFieldNames();
-        $aExpected = [
-            'apiUrl',
-            'maid',
-            'descriptor',
-            'additionalInfo',
-            'deleteCanceledOrder',
-            'deleteFailedOrder',
-        ];
-        $this->assertEquals($aExpected, $aPublicFieldNames, '', 0.0, 1, true);
     }
 }
